Update createPitchDeck; Add validatePitchDeckOwner
Before a user can upload a pitch deck, there a series of validation
steps that must occur:
1. They need to be logged in
2. They need to be authorized (i.e. in the right role, e.g. Founder)
3. The authorization payload must contain a valid ObjectId
4. The ObjectId must relate to an existing user in the database
5. The user needs to be within their grace period for uploading a new
   pitch deck

(1) and (2) are taken care of by the auth.required and grantAccess
middleware. (3) and (4) were already in the createUser method, but
needed to be moved to BEFORE the upload happens. Finally (5) needed to
be added before upload and after (3) and (4) were checked. These checks
have been placed in validatePitchDeckOwner, and createPitchDeck has been
updated to simply create the PitchDeck document once all checks have
been made and the file has been uploaded. In addition we (re)set the
user's pitch deck grace period for re-upload.